+++ danowski/apps/journals/migrations/0004_journal_slug.py
@@ -12,9 +12,6 @@ def set_slug_from_title(apps, schema_editor):
         j.slug = slugify(j.title)
         j.save()
 
-# NOTE: in django 1.8 use RunPython.noop
-def noop(apps, schema_editor):
-    pass
 
 class Migration(migrations.Migration):
 
@@ -29,7 +26,7 @@ class Migration(migrations.Migration):
             field=models.SlugField(blank=True),
             preserve_default=False,
         ),
-        migrations.RunPython(set_slug_from_title, noop),
+        migrations.RunPython(set_slug_from_title, migrations.RunPython.noop),
         # then set slug to be unique
         migrations.AlterField(
             model_name='journal',